% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Annual}
\alias{Annual}
\alias{Seasonal}
\alias{Harm}
\alias{Hyear}
\title{Construct basis vectors for Harmonic terms in a model}
\usage{
Annual(day, k = 4)

Seasonal(dates, k = 4)

Harm(theta, k = 4)

Hyear(x, k = 4)
}
\arguments{
\item{day, x}{time variable, assumed to be in elapsed days from a fixed origin}

\item{k}{the number of sine and cosine terms}

\item{dates}{time variable, as a Date object or can be coerced to Date}

\item{theta}{angle expressed in radians}
}
\value{
A matrix of harmonc terms with \code{length(days)} rows and \code{2*k} columns
}
\description{
Convenience function for specifying harmonic terms in a model
The two functions \code{Annual} and \code{Hyear} do the same job
but are retained for compatibility purposes.
}
\examples{
head(Annual(0:364))
}
