% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WWR_Software.R
\name{booktabs}
\alias{booktabs}
\alias{booktabs.default}
\alias{booktabs.noquote}
\alias{booktabs.xtable}
\alias{print.booktabs}
\title{Formatting function for tables in Rnw files}
\usage{
booktabs(x, ...)

\method{booktabs}{default}(x, align = c("@{}l", rep("r", ncol(x) - 1), "r@{}"), ...)

\method{booktabs}{noquote}(x, align = c("@{}l", rep("r", ncol(x) - 1), "r@{}"), ...)

\method{booktabs}{xtable}(x, ...)

\method{print}{booktabs}(
  x,
  floating = FALSE,
  hline.after = NULL,
  add.to.row = list(pos = list(-1, 0, nrow(x)), command = c("\\\\toprule ",
    "\\\\midrule ", "\\\\bottomrule ")),
  include.rownames = getOption("xtable.include.rownames", TRUE),
  ...
)
}
\arguments{
\item{x, align}{as for \code{\link[xtable]{xtable}}}

\item{...}{additional arguments passed on to methods}

\item{floating, hline.after, add.to.row, include.rownames}{as for \code{\link[xtable]{print.xtable}}}
}
\value{
A data frame, giving a suitably formatted LaTeX table when printed
}
\description{
Convenience function for drafting the course materia.s
}
\examples{
booktabs(head(iris))
}
