% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{cc}
\alias{cc}
\title{cc}
\format{
A data frame with 2000 rows and 2 columns:
\describe{
    \item{Longitude}{numeric: ... }
    \item{Latitude}{numeric: ... }
}
}
\usage{
cc
}
\description{
cc
}
\keyword{datasets}
