% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WWR_Software.R
\name{cs}
\alias{cs}
\title{Unquoted concatenation}
\usage{
cs(...)
}
\arguments{
\item{...}{strings, may be unquoted if standard R name format}
}
\value{
A string vector
}
\description{
Convenience function for inputing strings without quote signs
(unless the strings themselves are not standard R names).
}
\details{
Based on an original function \code{Cs} in the \code{Hmisc} package.
}
\examples{
(rhyme <- cs(Mary, had, a, little, lamb, "4 Christmas", 1:4))
}
