% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{data4}
\alias{data4}
\title{data4}
\format{
A data frame with 4 rows and 7 columns:
\describe{
    \item{Longitude}{numeric: ... }
    \item{Latitude}{numeric: ... }
    \item{Place}{factor: ... }
    \item{Sea}{numeric: ... }
    \item{Coast}{numeric: ... }
    \item{Depth}{array: ... }
    \item{Mud}{array: ... }
}
}
\usage{
data4
}
\description{
data4
}
\keyword{datasets}
