% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Self_starters.R
\name{fix_deriv}
\alias{fix_deriv}
\title{Modify Self-starting Model Functions}
\usage{
fix_deriv(func)
}
\arguments{
\item{func}{A function object, the output from \code{deriv} or \code{selfStart}}
}
\value{
A modified function with one extra line of code inserted
}
\description{
This is a work-around function to modify the output functions from \code{deriv}
or \code{selfStart} to allow more flexibility when using them with \code{nlmer}
from the \code{lme4} package.
}
\examples{
mod_fun <- deriv(~ b*v/(w - c), namevec = c("b", "c"), function.arg = function(v, w, b, c) {})
mod_fun_fix <- fix_deriv(mod_fun)
}
