% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{FSemiData4}
\alias{FSemiData4}
\title{FSemiData4}
\format{
A data frame with 1460 rows and 9 columns:
\describe{
    \item{Longitude}{numeric: ... }
    \item{Latitude}{numeric: ... }
    \item{Place}{factor: ... }
    \item{Sea}{numeric: ... }
    \item{Coast}{numeric: ... }
    \item{Depth}{array: ... }
    \item{Mud}{array: ... }
    \item{DayOfYear}{integer: ... }
    \item{Fsemi}{array: ... }
}
}
\usage{
FSemiData4
}
\description{
FSemiData4
}
\keyword{datasets}
