% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{GladstoneBream}
\alias{GladstoneBream}
\title{GladstoneBream}
\format{
A data frame with 337 rows and 23 columns:
\describe{
    \item{Trip}{character: ... }
    \item{ZoneNo}{integer: ... }
    \item{Zone}{factor: ... }
    \item{SiteId}{integer: ... }
    \item{SiteF}{factor: ... }
    \item{Site}{factor: ... }
    \item{Date}{Date: ... }
    \item{Year}{factor: ... }
    \item{Month}{factor: ... }
    \item{Tidal}{factor: ... }
    \item{Depth}{numeric: ... }
    \item{Sand}{numeric: ... }
    \item{Mud}{numeric: ... }
    \item{Gravel}{numeric: ... }
    \item{Rock}{numeric: ... }
    \item{Long}{numeric: ... }
    \item{Lat}{numeric: ... }
    \item{Casts}{numeric: ... }
    \item{Pikey}{integer: ... }
    \item{Yellowfin}{integer: ... }
    \item{Bream}{integer: ... }
    \item{P1}{numeric: ... }
    \item{P2}{numeric: ... }
}
}
\usage{
GladstoneBream
}
\description{
GladstoneBream
}
\keyword{datasets}
