% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{\%star\%}
\alias{\%star\%}
\title{Column product matrix}
\usage{
X \%star\% Y
}
\arguments{
\item{X, Y}{Numeric matrices with the same number of rows}
}
\value{
A matrix of products with \code{nrow(X)} rows and \code{ncol(X)*ncol(Y)} columns
}
\description{
Forms all pairwise products of the columns of X and Y suitable for interaction
terms in linear modeld and their allies
}
