% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Headrope}
\alias{Headrope}
\title{Headrope}
\format{
A data frame with 8594 rows and 13 columns:
\describe{
    \item{YearF}{factor: ... }
    \item{Y2K}{integer: ... }
    \item{Stock}{factor: ... }
    \item{Vessel}{factor: ... }
    \item{Days}{integer: ... }
    \item{Head}{numeric: ... }
    \item{Hull}{integer: ... }
    \item{Power}{numeric: ... }
    \item{Catch}{numeric: ... }
    \item{Banana}{numeric: ... }
    \item{Tiger}{numeric: ... }
    \item{Endeavour}{numeric: ... }
    \item{King}{numeric: ... }
}
}
\usage{
Headrope
}
\description{
Headrope
}
\keyword{datasets}
