% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Self_starters.R
\name{initial_value}
\alias{initial_value}
\alias{gompertz2_init}
\alias{gompertz3_init}
\alias{negexp_init}
\title{Initial Value Functions}
\usage{
gompertz2_init(x, y)

gompertz3_init(x, y)

negexp_init(x, y)
}
\arguments{
\item{x}{numeric vector; the independent variable}

\item{y}{numeric vector; the response variable}
}
\value{
A vector of initial values for the parameters with appropriate names
}
\description{
Free-standing functions to arrive at suitable initial values for non-linear model
fitting functions
}
\examples{
library(WWRData)
with(muscle, gompertz2_init(Conc, Length))
}
