% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WWR_Software.R
\name{Intersect}
\alias{Intersect}
\alias{Union}
\title{Extended union and intersect functions}
\usage{
Intersect(...)

Union(...)
}
\arguments{
\item{...}{One or more vectors}
}
\value{
A vector with the result of the operation
}
\description{
Like \code{\link[base]{intersect}} and \code{\link[base]{union}} but
allowing more than two arguments.
}
\examples{
Union(letters[1:3], 1:3, LETTERS[1:3], NULL, letters[1:5])
Intersect(letters, letters[1:12], letters[7:15])
}
