% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{isubset}
\alias{isubset}
\title{Generate all subsets}
\usage{
isubset(n, ..., chunks, chunkSize)
}
\arguments{
\item{n}{The size of the set}

\item{...}{currently ignred}

\item{chunks}{number of chunks to be supplied by the iterator}

\item{chunkSize}{size of a chunk (an alternative to \code{chunks})}
}
\value{
An iterator function for the purpose
}
\description{
An iterator to generate all subset of a collection
}
\examples{
get_subs <- isubset(5, chunks = 4)
get_subs$nextElem()
get_subs$nextElem()
get_subs$nextElem()
get_subs$nextElem()
get_subs$nextElem()
}
