% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{LB2004}
\alias{LB2004}
\title{LB2004}
\format{
A data frame with 11779 rows and 14 columns:
\describe{
    \item{Date}{Date: ... }
    \item{Year}{integer: ... }
    \item{Month}{factor: ... }
    \item{Week}{factor: ... }
    \item{Vcode}{factor: ... }
    \item{Latitude}{numeric: ... }
    \item{Longitude}{numeric: ... }
    \item{StockTig}{factor: ... }
    \item{StockEnde}{factor: ... }
    \item{StockEnsi}{factor: ... }
    \item{StockMerg}{factor: ... }
    \item{StockIndi}{factor: ... }
    \item{StockKing}{factor: ... }
    \item{Season}{factor: ... }
}
}
\usage{
LB2004
}
\description{
LB2004
}
\keyword{datasets}
