% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Locz}
\alias{Locz}
\title{Locz}
\format{
A data frame with  rows and  columns:
\describe{
    \item{}{complex: ... }
    \item{}{complex: ... }
    \item{}{complex: ... }
    \item{}{complex: ... }
}
}
\usage{
Locz
}
\description{
Locz
}
\keyword{datasets}
