% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{mostFreq}
\alias{mostFreq}
\alias{mostFreq.character}
\alias{mostFreq.factor}
\alias{mostFreq.logical}
\alias{mostFreq.numeric}
\title{Extract the most typical value from a vector}
\usage{
mostFreq(x, ...)

\method{mostFreq}{character}(x, ...)

\method{mostFreq}{factor}(x, ...)

\method{mostFreq}{logical}(x, ...)

\method{mostFreq}{numeric}(x, ...)
}
\arguments{
\item{x}{A suitable vector}

\item{...}{currently ignored}
}
\value{
The "most typical" value as described
}
\description{
Given a factor, character string vector, numeric vector or logical vector
extract a single "typical" value.  Either the median or the most frequently
occurring value
}
\examples{
data.frame(lapply(quine, mostFreq))
}
