% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{NPF0}
\alias{NPF0}
\title{Reasonably high quality map of the Northern Prawn Fishery}
\usage{
NPF0(col = alpha("#A67A5C", 0.5), axes = FALSE, las = 1)
}
\arguments{
\item{col}{colour for the mainland}

\item{axes}{logical: do you want axes?}

\item{las}{graphics parameter}
}
\value{
as for \code{\link[graphics]{lines}}
}
\description{
Demonstration function
}
\examples{
NPF0()
text(Latitude ~ Longitude, roundTrip, Locality, col = "red", xpd = NA, pos = 2)
}
