% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{NPFBigCoastline}
\alias{NPFBigCoastline}
\title{NPFBigCoastline}
\format{
A data frame with 210736 rows and 4 columns:
\describe{
    \item{x}{numeric: ... }
    \item{y}{numeric: ... }
    \item{Longitude}{numeric: ... }
    \item{Latitude}{numeric: ... }
}
}
\usage{
NPFBigCoastline
}
\description{
NPFBigCoastline
}
\examples{
if("WWRGraphics" \%in\% loadedNamespaces()) {
plot(Latitude ~ Longitude, NPFBigCoastline,
     asp = 1, type = "l", col = alpha("dark green", 0.5))
points(Latitude ~ Longitude, roundTrip)
text(Latitude ~ Longitude, roundTrip, Locality,
     pos = avoid(Longitude, Latitude), xpd=NA)
} else cat("Attach WWRGraphics package first\n")
}
\keyword{datasets}
