% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{oneSERule}
\alias{oneSERule}
\alias{oneSERule.rpart}
\title{One Standard Error Rule}
\usage{
oneSERule(tree, f, ...)

\method{oneSERule}{rpart}(tree, f = 1, ...)
}
\arguments{
\item{tree}{A tree object fitted with \code{\link[rpart]{rpart}}}

\item{f}{The number of standard errors (if not the default 1)}

\item{...}{Presently ignored}
}
\value{
A value suitable for specifying as the cp argument of \code{\link[rpart]{prune}}
}
\description{
Extract the Cp value needed to implement the 1-standard-error rule in
tree prunng for tree objexts
}
