% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{OOB}
\alias{OOB}
\alias{OOB.default}
\alias{OOB.randomForest}
\title{Extract the out-of-bag error rate from a randomForest object}
\usage{
OOB(obj, ...)

\method{OOB}{default}(obj, ...)

\method{OOB}{randomForest}(obj, ...)
}
\arguments{
\item{obj}{A classification object produced by \code{\link[randomForest]{randomForest}}}

\item{...}{Presently ignored}
}
\value{
The overall out-of-bag error rate, if available
}
\description{
Convenience function
}
