% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfImputeUnsupervised.R
\name{rfImputeUnsupervised}
\alias{rfImputeUnsupervised}
\alias{rfImputeUnsupervised.default}
\alias{rfImputeUnsupervised.formula}
\title{Missing value imputation}
\usage{
rfImputeUnsupervised(object, ...)

\method{rfImputeUnsupervised}{default}(object, iter = 5, ...)

\method{rfImputeUnsupervised}{formula}(object, data = environment(object), ...)
}
\arguments{
\item{object}{A one-sided formula, or a data frame containing missing values in some of the variables.  Missing values in \code{Date} or \code{POSIXt} objects are not accommodated}

\item{...}{additional arguments passed on to \code{randomForest}}

\item{data}{A data frame used when \code{x} is a \code{formula}}
}
\value{
A data frame like \code{object} with missing values imputed, if possible
}
\description{
An unsupervised missing value imputation method
using \code{randomForest} methods.  Somewhat experimental
}
