% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session}
\alias{session}
\alias{print.session}
\title{Get the session script}
\usage{
session(n, scripts = "scripts", pdf = "PDF", here = getwd())

\method{print}{session}(x, ...)
}
\arguments{
\item{n}{integer: the session number}

\item{scripts}{the sub-folder containint the scripts, .R files}

\item{pdf}{the sub-folder containint the pdf files, .pdf}

\item{here}{where the results should go, normally the working directory}

\item{x}{An object of class "session"}

\item{...}{extra arguments (currently ignored)}
}
\value{
A character string vector which when printed yields
        the required pared-down script file.
}
\description{
A command-line (console) tool to fetch the script for a
particular session, remove some unnecessary code and place
a copy in a file in the main working directory.
}
\details{
It also places a copy of the display pdf file in the
main working directory as well.
}
\examples{
\dontrun{
session(1)
}
}
