% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{ShortCC}
\alias{ShortCC}
\title{ShortCC}
\format{
A data frame with 2001 rows and 2 columns:
\describe{
    \item{x}{numeric: ... }
    \item{y}{numeric: ... }
}
}
\usage{
ShortCC
}
\description{
Coordinates defining a curve through the main
fishing areas of the NPF
}
\keyword{datasets}
