% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Self_starters.R
\name{SSgompertz2}
\alias{SSgompertz2}
\alias{SSgompertz3}
\title{Self-starting Model Functions}
\usage{
SSgompertz2(x, b0, b1, rho)

SSgompertz3(x, b0, b1, rho)
}
\arguments{
\item{x}{numeric vector, time variable or analogue}

\item{b0, b1.rho}{place-holders for the parameter estimates.}
}
\value{
An evaluated model function, with a derivative matrix attribute
}
\description{
Self-starting model fitting functions for non-linear Gompertz Growth Curve models.
The non-linear model is of the form ~ exp(b0 - b1*rho^x)
}
\details{
The functions are not used directly, but in non-linear model formulae
}
\examples{
fm <- nls(Length ~ SSgompertz2(Conc, b0, b1, rho), data = muscle)
coef(fm)
}
