% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Self_starters.R
\name{SSnegexp}
\alias{SSnegexp}
\title{Negative Exponential Model}
\usage{
SSnegexp(t, b0, b1, theta)
}
\arguments{
\item{t}{The time variable, or analogue}

\item{b0, b1, theta}{place-holders for the parameter estimates}
}
\value{
An evaluated model function, with a derivative matrix attribute
}
\description{
A self-starting model fitting function for a negative exponential model.
The model formula is ~ b0 + b1*exp(-t/theta)
}
\examples{
fm <- nls(Weight ~ SSnegexp(Days, b0, b1, theta), wtloss)
coef(fm)
}
