% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WWR_Software.R
\name{subset.data.frame}
\alias{subset.data.frame}
\title{Extended subset function}
\usage{
\method{subset}{data.frame}(x, subset, select, drop = FALSE, ...)
}
\arguments{
\item{x, subset, select, drop}{As for \code{\link[base]{subset}}}

\item{...}{additional arguments passed on}
}
\value{
A data frame, as for \code{\link[base]{subset}}
}
\description{
Like \code{\link[base]{subset}} but allowing any form of indexing for
the \code{subset} argument, not just logical.
}
\examples{
subset(iris, 1:10)
}
