% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{teff}
\alias{teff}
\title{teff}
\format{
A data frame with 7029 rows and 3 columns:
\describe{
    \item{Longitude}{numeric: ... }
    \item{Latitude}{numeric: ... }
    \item{TigerE}{numeric: ... }
}
}
\usage{
teff
}
\description{
teff
}
\keyword{datasets}
