% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Tigers}
\alias{Tigers}
\title{Tigers}
\format{
A data frame with 12569 rows and 13 columns:
\describe{
    \item{Survey}{factor: ... }
    \item{Date}{factor: ... }
    \item{Longitude}{numeric: ... }
    \item{Latitude}{numeric: ... }
    \item{Pesc}{numeric: ... }
    \item{Psem}{numeric: ... }
    \item{Total}{numeric: ... }
    \item{Sea}{numeric: ... }
    \item{Coast}{numeric: ... }
    \item{Depth}{numeric: ... }
    \item{Mud}{numeric: ... }
    \item{DayOfYear}{integer: ... }
    \item{ElapsedDays}{integer: ... }
}
}
\usage{
Tigers
}
\description{
Tigers
}
\keyword{datasets}
