% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{today}
\alias{today}
\alias{right_now}
\title{Time stamping functions}
\usage{
today()

right_now()
}
\value{
a character string giving the date or the precise time
}
\description{
Convenience functions for time-stamping output or graphics
}
\examples{
hist(~log(crim), Boston, binWidth(0.5), fill = pal_desert)
text("top right", right_now(), cex = 0.7, col = "steel blue")
}
