% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{twoWay}
\alias{twoWay}
\title{Special interaction function}
\usage{
twoWay(day, sea, k = c(3, 2))
}
\arguments{
\item{day}{Day of year}

\item{sea}{Distance out to sea}

\item{k}{integer: two components}
}
\value{
A matix for use in model specification
}
\description{
Convenience function, mainly as a demonstration.
Used in WWR examples
}
