% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WWR_Software.R
\name{untibble}
\alias{untibble}
\title{Turn off tible production}
\usage{
untibble(tibble)
}
\arguments{
\item{tibble}{Any object that can be coerced to a data frame, including tibbles}
}
\value{
An honest data frame
}
\description{
Convenience function to turn a \code{"tibble"} back into an
honest data frame
}
