% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WWR_Software.R
\name{use_packages}
\alias{use_packages}
\title{Attach multiple packages}
\usage{
use_packages(..., quietly = FALSE)
}
\arguments{
\item{...}{package names, unquoted}
}
\value{
a list of packages attached to the search path
}
\description{
Attach a series of packages to the search path, optionally
installing them if needed
}
\examples{
s0 <- search()
use_packages(mgcv, lme4, splines, quietly = FALSE)
setdiff(search(), s0)
}
