% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WWR_Software.R
\name{zfill}
\alias{zfill}
\alias{fill0}
\title{Zero fill strings}
\usage{
zfill(x)

fill0(x)
}
\arguments{
\item{x}{A vector which will be coerced to charater}
}
\value{
A zero-padded character string vector
}
\description{
Pads out strings with leading zeros to a uniform character size.
\code{fill0} does the same job using a different algorighm
}
\examples{
zfill(1:10)
fill0(cs(Mary, had, a, little, lamb))
}
